refactor: controller 피드백 반영
