+++ src/main/java/wooteco/chess/service/ChessGameService.java
@@ -3,8 +3,8 @@ package wooteco.chess.service;
 import org.springframework.stereotype.Service;
 import wooteco.chess.domain.board.Board;
 import wooteco.chess.domain.board.BoardFactory;
-import wooteco.chess.domain.dao.BoardDAO;
-import wooteco.chess.domain.dao.TurnDAO;
+import wooteco.chess.dao.BoardDAO;
+import wooteco.chess.dao.TurnDAO;
 import wooteco.chess.domain.piece.Piece;
 import wooteco.chess.domain.piece.Team;
 import wooteco.chess.domain.position.Position;
